chore(battery): move out led logic from app.c
