+++ apps/battery-monitor/include/led.h
@@ -1,0 +1,28 @@
+#ifndef LED_H
+#define LED_H
+
+#ifdef __cplusplus
+extern "C" {
+#endif
+
+typedef enum {
+  NONE = 0,
+  RED,
+  GREEN,
+  ORANGE,
+} led_color_t;
+
+// Naming based on power board schematic
+typedef enum {
+  LED6 = 0,
+  LED7,
+} led_t;
+
+void blink_leds_red(void);
+void set_led_color(led_t led, led_color_t color);
+
+#ifdef __cplusplus
+}
+#endif
+
+#endif /* LED_H */
